tests/requirements.txt
coincidence>=0.1.0
coverage>=5.1
coverage-pyver-pragma>=0.2.1
domdf-python-tools[testing]>=2.0.1
gitpython>=3.1.14
iniconfig!=1.1.0,>=1.0.1
pytest>=6.0.0
pytest-cov>=2.8.1
pytest-randomly>=3.3.1
pytest-timeout>=1.4.2
